\name{descriptor}
\alias{descriptor}

\title{
Distribution of Assays as percentage of cell lines
}
\description{

descriptor() produces panels of boxplots displaying the
distribution of various factors(e.g. ChIP-seq of H3k9ac
histone modification) in percentage of cell lines. It asks
for user input C1 or C2  and plots the following boxplots.

choice = 1 output is collection of boxplots explaining the
distribution of percentage of cell lines having or overlapping
with the input factors (e.g. H3k9ac).

choice = 2 displays the distribution of percentage of cell
lines having or overlapping with increasing number of factors
for the selected group (e.g. WE or RE etc).
}
\usage{

descriptor(name, factors, groups,
choice, choice2group = NULL, title = NULL)
}

\arguments{
\item{name}{
Name of the "OMICs" object containing the assay data
}
\item{factors}{
A vector containing the name of the Assays read by "InputOMICs()"
}
\item{groups}{
Name of all or subset of groups crated by "create_group()"
}

\item{choice}{
Numerical value between 1 and 2. See Description and vignettes for
details of plots generated by this 2 choices
}

\item{choice2group}{
Used when choice is 2. This argument allows a name of group created
by "create_group"
}

\item{title}{
Title of the plot
}
}
\value{
Displays various plots as per the combination of various input arguments.
}
\author{
Subhadeep Das <subhadeep1024@gmail.com>
}

\examples{

Assays <- c("H2az_Cell_wise", "H3k9ac_Cell_wise",
"H3k4me1_Cell_wise", "CAGE_Cell_wise_v2")

#OMICsdata <- OMICs()
#path = system.file("extdata/data", package = "OMICsPCAdata")


#InputOMICs(name = "OMICsdata", Assays = Assays, typeisfile = TRUE,
#path = path)

#create_group(name = "OMICsdata",
#group_names = c("WE","RE","NE","IntE"),
#grouping_factor = "CAGE_Cell_wise_v2",
#comparison = c(">=","\%in\%","==","\%in\%"),
#condition = c("25","1:5","0","6:24"))
#
#
#descriptor(name = "OMICsdata", factors = c("H2az_Cell_wise",
#"H3k4me1_Cell_wise"),groups = c("WE","RE", "IntE","NE"),
#choice = 1, title = "distribution of percentage of cell types")

#descriptor(name = "OMICsdata", factors = c("H2az_Cell_wise",
#"H3k4me1_Cell_wise"), groups = c("WE","RE", "IntE","NE"),
#choice = 2,
#choice2group = "WE")

}
