\name{OMICs-class}
\Rdversion{1.1}
\docType{class}
\alias{OMICs-class}

\title{Class "OMICs"}
\description{
An S4 class designed to store data generated by the
functions of package "OMICsPCA"
}
\section{Objects from the Class}{
Objects can be created by calls of the form new("OMICs", ...).
}
\section{Slots}{
\describe{
\item{InputAssays}{Object of class"list"}
\item{groups}{Object of class "list"}
\item{PCA}{Object of class "list"}
\item{integratedPCA}{Object of class "list"}
}
}

\author{
Subhadeep Das
}

\examples{
showClass("OMICs")
}

\keyword{classes}
